refactor buf
