+++ crabml-core/src/tensor/cpu/mod.rs
@@ -1,6 +1,5 @@
 pub mod arithmetic;
 pub mod buf;
-pub mod quant;
 pub mod tensor;
 pub mod validation;
